cleanup: module organization, shared git check, source round-trip
- Move restore_overlays, update_overlays, and switch_overlay out of
  create.rs into src/restore.rs, src/update.rs, and src/switch.rs;
  create.rs now only holds overlay-creation logic. The old update.rs
  (version-string/update-check code) becomes src/version.rs.
- Replace hand-rolled .git checks in edit, sync, and create with
  validate_git_repo, unifying the error text.
- Add OverlaySource::reapply_reference (TDD) encoding the
  source-to-reference round-trip, replacing the two matches in
  restore_overlays that rebuilt references far from the type.

Fixes #371
