+++ src/create.rs
@@ -1,412 +1,10 @@
 use anyhow::{Context, Result, bail};
 use colored::Colorize;
-use log::debug;
 use std::fs;
 use std::path::{Path, PathBuf};
 
-use crate::ApplyOptions;
-use crate::ConflictStrategy;
-use crate::OverlayName;
-use crate::apply_overlay;
-use crate::cache::CacheManager;
-use crate::canonicalize_path;
 use crate::detection;
-use crate::github::GitHubSource;
-use crate::remove_overlay;
-use crate::remove_single_overlay;
 use crate::selection;
-use crate::state::{
-    OVERLAYS_DIR, OverlaySource, STATE_DIR, SourceResolver, list_applied_overlays,
-    load_external_states, load_overlay_state, normalize_overlay_name,
-};
-use crate::validate_git_repo;
-
-/// Restore overlays after git clean or other removal.
-///
-/// Uses external state backup (`~/.local/share/repoverlay/applied/`) to recover
-/// overlays that were removed by `git clean -fdx` or similar operations.
-///
-/// # Workflow
-///
-/// 1. Load external state backup for the target repository
-/// 2. For each saved overlay state, re-apply using original source
-pub(crate) fn restore_overlays(
-    target: &Path,
-    dry_run: bool,
-    conflict_strategy: ConflictStrategy,
-    merge: bool,
-) -> Result<()> {
-    debug!(
-        "restore_overlays: target={}, dry_run={}, conflict_strategy={:?}",
-        target.display(),
-        dry_run,
-        conflict_strategy
-    );
-    let target = canonicalize_path(target, "Target directory")?;
-    validate_git_repo(&target)?;
-
-    // Load external state
-    let external_states = load_external_states(&target)?;
-    debug!("found {} external states to restore", external_states.len());
-
-    let restorable_profiles = crate::profile_plan::list_restorable_profiles(&target)?;
-    debug!("found {} profile(s) to restore", restorable_profiles.len());
-
-    if external_states.is_empty() && restorable_profiles.is_empty() {
-        println!("{} Nothing to restore.", "Status:".bold());
-        println!("  No external backup found for this repository.");
-        return Ok(());
-    }
-
-    if !external_states.is_empty() {
-        println!(
-            "{} {} overlay(s) to restore:",
-            "Found".blue().bold(),
-            external_states.len()
-        );
-
-        for state in &external_states {
-            println!("  - {}", state.name);
-            match &state.source {
-                OverlaySource::Local { path, .. } => {
-                    println!("    Source: {}", path.display());
-                }
-                OverlaySource::GitHub { url, git_ref, .. } => {
-                    println!("    Source: {url} ({git_ref})");
-                }
-                OverlaySource::OverlayRepo {
-                    org,
-                    repo,
-                    name: overlay_name,
-                    ..
-                } => {
-                    println!("    Source: {org}/{repo}/{overlay_name} (overlay repo)");
-                }
-                OverlaySource::Library { name } => {
-                    println!("    Source: {name} (library)");
-                }
-            }
-        }
-    }
-
-    if !restorable_profiles.is_empty() {
-        println!(
-            "{} {} profile(s) to restore:",
-            "Found".blue().bold(),
-            restorable_profiles.len()
-        );
-        for profile in &restorable_profiles {
-            println!("  - {} ({})", profile.name, profile.harness);
-        }
-    }
-
-    if dry_run {
-        println!("\n{} Dry run - no changes made.", "Note:".yellow());
-        return Ok(());
-    }
-
-    println!();
-
-    // Restore each overlay
-    let mut restored = Vec::new();
-    let mut failures = Vec::new();
-
-    for state in external_states {
-        let source_str = match &state.source {
-            OverlaySource::Local { path, .. } => path.to_string_lossy().to_string(),
-            OverlaySource::GitHub {
-                url,
-                owner,
-                repo,
-                subpath,
-                ..
-            } => {
-                // Reconstruct URL with subpath so resolve_source can find the
-                // specific overlay instead of falling into browse/selection mode.
-                subpath.as_ref().map_or_else(
-                    || url.clone(),
-                    |subpath| format!("https://github.com/{owner}/{repo}/tree/HEAD/{subpath}"),
-                )
-            }
-            OverlaySource::OverlayRepo {
-                org,
-                repo,
-                name: overlay_name,
-                ..
-            } => {
-                format!("{org}/{repo}/{overlay_name}")
-            }
-            OverlaySource::Library { name } => name.clone(),
-        };
-
-        let ref_override = match &state.source {
-            OverlaySource::GitHub { git_ref, .. } => Some(git_ref.clone()),
-            OverlaySource::Local { .. }
-            | OverlaySource::Library { .. }
-            | OverlaySource::OverlayRepo { .. } => None,
-        };
-
-        // Re-apply the overlay. Always use Force since restore's purpose is to
-        // re-create missing/broken symlinks from external backup state.
-        match apply_overlay(
-            &source_str,
-            &target,
-            &ApplyOptions {
-                name_override: Some(state.name.clone()),
-                ref_override,
-                update_cache: true,
-                conflict_strategy: ConflictStrategy::Force,
-                merge,
-                ..ApplyOptions::default()
-            },
-        ) {
-            Ok(()) => {
-                println!("  {} Restored '{}'", "✓".green(), state.name);
-                restored.push(state.name.clone());
-            }
-            Err(e) => {
-                let error = e.to_string();
-                eprintln!(
-                    "  {} Failed to restore '{}': {}",
-                    "Error:".red(),
-                    state.name,
-                    error
-                );
-                failures.push((state.name.clone(), error));
-            }
-        }
-    }
-
-    // Profiles are restored after overlays so any overlay files a profile
-    // depends on are already back in place.
-    if !restorable_profiles.is_empty() {
-        match crate::profile_plan::restore_profiles(&target) {
-            Ok(count) => {
-                println!(
-                    "  {} Restored {count} profile(s).",
-                    "Profiles:".cyan().bold()
-                );
-            }
-            Err(e) => {
-                eprintln!("  {} Failed to restore profiles: {}", "Error:".red(), e);
-            }
-        }
-    }
-
-    println!();
-    println!("{}", "Restore summary:".bold());
-    if restored.is_empty() {
-        println!("  Restored: none");
-    } else {
-        println!("  Restored: {}", restored.join(", "));
-    }
-
-    if failures.is_empty() {
-        println!("  Failed: none");
-    } else {
-        println!("  Failed:");
-        for (name, error) in &failures {
-            println!("    - {name}: {error}");
-        }
-
-        let failed_names = failures
-            .iter()
-            .map(|(name, _)| name.as_str())
-            .collect::<Vec<_>>()
-            .join(", ");
-        let restored_names = if restored.is_empty() {
-            "none".to_string()
-        } else {
-            restored.join(", ")
-        };
-        bail!(
-            "Failed to restore {} overlay(s): {failed_names}. Restored: {restored_names}",
-            failures.len()
-        );
-    }
-
-    Ok(())
-}
-
-/// Update applied overlays from remote sources.
-///
-/// Only GitHub-sourced overlays can be updated. Local overlays are skipped.
-///
-/// # Workflow
-///
-/// 1. List applied overlays (optionally filtered by name)
-/// 2. For each GitHub overlay, check remote for new commits
-/// 3. Report available updates
-/// 4. If not dry-run, remove and re-apply each overlay with updated cache
-#[allow(clippy::needless_pass_by_value)]
-pub(crate) fn update_overlays(
-    target: &Path,
-    name: Option<String>,
-    dry_run: bool,
-    conflict_strategy: ConflictStrategy,
-    merge: bool,
-) -> Result<()> {
-    debug!(
-        "update_overlays: target={}, name={:?}, dry_run={}, conflict_strategy={:?}",
-        target.display(),
-        name,
-        dry_run,
-        conflict_strategy
-    );
-    let target = canonicalize_path(target, "Target directory")?;
-    let overlays_dir = target.join(STATE_DIR).join(OVERLAYS_DIR);
-
-    if !overlays_dir.exists() {
-        bail!("No overlays are currently applied in: {}", target.display());
-    }
-
-    let applied_overlays = list_applied_overlays(&target)?;
-    if applied_overlays.is_empty() {
-        bail!("No overlays are currently applied in: {}", target.display());
-    }
-
-    // Filter to just the specified overlay if name provided
-    let overlays_to_check: Vec<OverlayName> = if let Some(ref name) = name {
-        let normalized = normalize_overlay_name(name)?;
-        if !applied_overlays.iter().any(|n| n == normalized.as_str()) {
-            let names: Vec<&str> = applied_overlays.iter().map(OverlayName::as_str).collect();
-            bail!(
-                "Overlay '{}' is not applied. Available: {}",
-                name,
-                names.join(", ")
-            );
-        }
-        vec![OverlayName::new(normalized)]
-    } else {
-        applied_overlays
-    };
-
-    let cache = CacheManager::new()?;
-    let mut updates_available = Vec::new();
-
-    // Check for updates
-    for overlay_name in &overlays_to_check {
-        let state = load_overlay_state(&target, overlay_name.as_str())?;
-
-        if let OverlaySource::GitHub {
-            owner,
-            repo,
-            git_ref,
-            commit,
-            subpath,
-            url,
-            ..
-        } = &state.source
-        {
-            let source = GitHubSource {
-                owner: owner.clone(),
-                repo: repo.clone(),
-                git_ref: git_ref.parse().map_err(|e: String| anyhow::anyhow!(e))?,
-                subpath: subpath.as_ref().map(PathBuf::from),
-            };
-
-            match cache.check_for_updates(&source) {
-                Ok(Some(new_commit)) => {
-                    updates_available.push((
-                        overlay_name.clone(),
-                        state.name.clone(),
-                        url.clone(),
-                        commit.clone(),
-                        new_commit,
-                    ));
-                }
-                Ok(None) => {
-                    println!("  {} {} is up to date", "✓".green(), state.name);
-                }
-                Err(e) => {
-                    println!(
-                        "  {} Could not check {} for updates: {}",
-                        "?".yellow(),
-                        state.name,
-                        e
-                    );
-                }
-            }
-        } else if state.source.is_library() {
-            // Library overlays are managed in-repo — update via git
-            println!(
-                "  {} {} (library overlay — update via git)",
-                "-".dimmed(),
-                state.name,
-            );
-        } else if state.source.is_updatable() {
-            // OverlayRepo sources: update by re-applying from the overlay repo
-            println!(
-                "  {} {} ({} source, update via 'repoverlay restore')",
-                "-".dimmed(),
-                state.name,
-                state.source.source_type_label()
-            );
-        } else {
-            println!(
-                "  {} {} is a {} overlay (not updatable)",
-                "-".dimmed(),
-                state.name,
-                state.source.source_type_label()
-            );
-        }
-    }
-
-    if updates_available.is_empty() {
-        println!("\n{} All overlays are up to date.", "Status:".bold());
-        return Ok(());
-    }
-
-    println!(
-        "\n{} {} update(s) available:",
-        "Found".blue().bold(),
-        updates_available.len()
-    );
-
-    for (_, name, url, old_commit, new_commit) in &updates_available {
-        println!("  {} {}", "↑".cyan(), name);
-        println!(
-            "    {}  →  {}",
-            &old_commit[..7.min(old_commit.len())],
-            &new_commit[..7.min(new_commit.len())]
-        );
-        println!("    {}", url.dimmed());
-    }
-
-    if dry_run {
-        println!("\n{} Dry run - no changes made.", "Note:".yellow());
-        return Ok(());
-    }
-
-    println!();
-
-    // Apply updates
-    for (normalized_name, _, _, _, _) in &updates_available {
-        let state = load_overlay_state(&target, normalized_name.as_str())?;
-
-        if let OverlaySource::GitHub { url, git_ref, .. } = &state.source {
-            // Remove old overlay
-            let overlays_dir = target.join(STATE_DIR).join(OVERLAYS_DIR);
-            remove_single_overlay(&target, &overlays_dir, normalized_name.as_str())?;
-
-            // Re-apply with update
-            apply_overlay(
-                url,
-                &target,
-                &ApplyOptions {
-                    name_override: Some(state.name.clone()),
-                    ref_override: Some(git_ref.clone()),
-                    update_cache: true,
-                    conflict_strategy,
-                    merge,
-                    ..ApplyOptions::default()
-                },
-            )?;
-        }
-    }
-
-    Ok(())
-}
 
 /// Expand glob patterns in include paths relative to the source directory.
 ///
@@ -827,35 +425,6 @@ pub(crate) fn create_overlay_with_files(
     Ok(())
 }
 
-/// Switch to a different overlay by removing all existing overlays first.
-///
-/// Atomic replacement of all overlays - useful for switching between different
-/// configurations (e.g., different AI agent setups).
-///
-/// # Workflow
-///
-/// 1. Remove all existing overlays (if any)
-/// 2. Apply the new overlay
-pub(crate) fn switch_overlay(source: &str, target: &Path, options: &ApplyOptions) -> Result<()> {
-    validate_git_repo(target)?;
-
-    // Check if any overlays are currently applied
-    let state_dir = target.join(STATE_DIR).join(OVERLAYS_DIR);
-    let has_overlays = state_dir.exists() && fs::read_dir(&state_dir)?.next().is_some();
-
-    if has_overlays {
-        println!("{} existing overlays...", "Removing".yellow().bold());
-        // Remove all existing overlays
-        remove_overlay(target, None, true, options.dry_run)?;
-    }
-
-    // Apply the new overlay
-    println!("{} new overlay...", "Applying".blue().bold());
-    apply_overlay(source, target, options)?;
-
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1040,323 +609,6 @@ mod tests {
             let config = generate_overlay_config("");
             assert!(config.contains("name = \n"));
             assert!(config.contains("overlay ="));
-        }
-    }
-
-    mod restore_overlays_tests {
-        use super::*;
-        use crate::state::{OverlayState, external_state_dir_for_target, load_external_states};
-        use crate::testutil::{TestContext, create_overlay_dir};
-
-        #[test]
-        fn does_not_restore_explicitly_removed_overlay() {
-            // This test verifies that `restore` does not re-apply overlays that were
-            // explicitly removed via `repoverlay remove`. The issue is that if external
-            // state exists but in-repo state was intentionally deleted (via remove command),
-            // restore should NOT re-apply the overlay.
-            //
-            // The fix marks external state with `removed_at` timestamp when an overlay
-            // is explicitly removed, so that `restore` knows to skip it.
-
-            let ctx = TestContext::new().with_overlay(&[(".envrc", "export FOO=bar")]);
-
-            // Use canonical path consistently (this is what restore_overlays does internally)
-            let canonical_repo_path = ctx.repo_path().canonicalize().unwrap();
-
-            // Step 1: Apply the overlay
-            apply_overlay(
-                ctx.overlay_source(),
-                ctx.repo_path(),
-                &ApplyOptions::default(),
-            )
-            .expect("apply should succeed");
-
-            // Verify overlay was applied
-            assert!(
-                ctx.file_exists(".envrc"),
-                "overlay file should exist after apply"
-            );
-            assert!(
-                ctx.overlay_state_exists("test-overlay") || ctx.state_dir_exists(),
-                "in-repo state should exist"
-            );
-
-            // Verify external state was saved (before removal)
-            let ext_dir = external_state_dir_for_target(&canonical_repo_path).unwrap();
-            assert!(ext_dir.exists(), "external state directory should exist");
-
-            // Step 2: Remove the overlay (this simulates explicit user removal)
-            // This should mark the external state with `removed_at` instead of deleting it.
-            let applied = list_applied_overlays(ctx.repo_path()).expect("list should work");
-            assert!(
-                !applied.is_empty(),
-                "at least one overlay should be applied"
-            );
-            let overlay_name = &applied[0];
-
-            remove_overlay(
-                ctx.repo_path(),
-                Some(overlay_name.to_string()),
-                false,
-                false,
-            )
-            .expect("remove should succeed");
-
-            // Verify overlay was removed from in-repo state
-            assert!(!ctx.file_exists(".envrc"), "overlay file should be removed");
-            assert!(
-                !ctx.overlay_state_exists(overlay_name.as_str()),
-                "in-repo state should be removed"
-            );
-
-            // Verify external state file still exists (with removed_at marker)
-            let ext_state_file = ext_dir.join(format!("{overlay_name}.ccl"));
-            assert!(
-                ext_state_file.exists(),
-                "external state file should still exist (as tombstone)"
-            );
-
-            // Read the external state and verify it has removed_at set
-            let content = fs::read_to_string(&ext_state_file).unwrap();
-            let ext_state: OverlayState = sickle::from_str(&content).unwrap();
-            assert!(
-                ext_state.removed_at.is_some(),
-                "external state should have removed_at marker"
-            );
-
-            // Verify load_external_states skips removed overlays
-            let external_states =
-                load_external_states(&canonical_repo_path).expect("load should work");
-            assert_eq!(
-                external_states.len(),
-                0,
-                "load_external_states should skip removed overlays"
-            );
-
-            // Step 3: Call restore - this SHOULD NOT restore the overlay
-            // because it was explicitly removed (has removed_at marker).
-            restore_overlays(ctx.repo_path(), false, ConflictStrategy::default(), false)
-                .expect("restore should succeed");
-
-            // Step 4: Verify the overlay was NOT restored
-            assert!(
-                !ctx.file_exists(".envrc"),
-                "overlay file should NOT be restored after explicit removal"
-            );
-        }
-
-        #[test]
-        fn restores_overlay_after_git_clean() {
-            // This test verifies that `restore` DOES re-apply overlays when
-            // in-repo state is missing due to `git clean -fdx` (not explicit removal).
-            //
-            // The external state should NOT have `removed_at` set because the
-            // overlay was not explicitly removed.
-
-            let ctx = TestContext::new().with_overlay(&[(".envrc", "export FOO=bar")]);
-            let canonical_repo_path = ctx.repo_path().canonicalize().unwrap();
-
-            // Step 1: Apply the overlay
-            apply_overlay(
-                ctx.overlay_source(),
-                ctx.repo_path(),
-                &ApplyOptions::default(),
-            )
-            .expect("apply should succeed");
-
-            let applied = list_applied_overlays(ctx.repo_path()).expect("list should work");
-            assert!(!applied.is_empty());
-
-            // Verify external state exists and doesn't have removed_at
-            let ext_states = load_external_states(&canonical_repo_path).unwrap();
-            assert_eq!(ext_states.len(), 1);
-            assert!(
-                ext_states[0].removed_at.is_none(),
-                "external state should NOT have removed_at"
-            );
-
-            // Step 2: Simulate `git clean -fdx` by removing only in-repo state
-            // This does NOT call remove_overlay, so external state stays intact.
-            fs::remove_dir_all(ctx.repo_path().join(".repoverlay")).unwrap();
-            // Also remove the overlay files (as git clean would)
-            fs::remove_file(ctx.repo_path().join(".envrc")).unwrap();
-
-            // Verify in-repo state is gone
-            assert!(!ctx.state_dir_exists(), "in-repo state should be removed");
-            assert!(
-                !ctx.file_exists(".envrc"),
-                "overlay files should be removed"
-            );
-
-            // External state should still be loadable (no removed_at marker)
-            let ext_states_after = load_external_states(&canonical_repo_path).unwrap();
-            assert_eq!(
-                ext_states_after.len(),
-                1,
-                "external state should still be loadable"
-            );
-
-            // Step 3: Call restore - this SHOULD restore the overlay
-            restore_overlays(ctx.repo_path(), false, ConflictStrategy::default(), false)
-                .expect("restore should succeed");
-
-            // Step 4: Verify the overlay WAS restored
-            assert!(
-                ctx.file_exists(".envrc"),
-                "overlay file should be restored after git clean"
-            );
-        }
-
-        #[test]
-        fn restores_all_overlays_after_git_clean_returns_ok() {
-            let ctx = TestContext::new();
-            let overlay_a = create_overlay_dir(&[(".envrc", "export FOO=bar")]);
-            let overlay_b = create_overlay_dir(&[(".toolrc", "tool = true")]);
-
-            apply_overlay(
-                overlay_a.path().to_str().unwrap(),
-                ctx.repo_path(),
-                &ApplyOptions {
-                    name_override: Some("overlay-a".to_string()),
-                    ..ApplyOptions::default()
-                },
-            )
-            .expect("first apply should succeed");
-            apply_overlay(
-                overlay_b.path().to_str().unwrap(),
-                ctx.repo_path(),
-                &ApplyOptions {
-                    name_override: Some("overlay-b".to_string()),
-                    ..ApplyOptions::default()
-                },
-            )
-            .expect("second apply should succeed");
-
-            fs::remove_dir_all(ctx.repo_path().join(".repoverlay")).unwrap();
-            fs::remove_file(ctx.repo_path().join(".envrc")).unwrap();
-            fs::remove_file(ctx.repo_path().join(".toolrc")).unwrap();
-
-            restore_overlays(ctx.repo_path(), false, ConflictStrategy::default(), false)
-                .expect("restore should succeed when every overlay restores");
-
-            assert!(
-                ctx.file_exists(".envrc"),
-                "first overlay should be restored"
-            );
-            assert!(
-                ctx.file_exists(".toolrc"),
-                "second overlay should be restored"
-            );
-        }
-
-        #[test]
-        fn returns_error_summary_after_partial_restore_failure() {
-            let ctx = TestContext::new();
-            let overlay_a = create_overlay_dir(&[(".envrc", "export FOO=bar")]);
-            let overlay_b = create_overlay_dir(&[(".toolrc", "tool = true")]);
-
-            apply_overlay(
-                overlay_a.path().to_str().unwrap(),
-                ctx.repo_path(),
-                &ApplyOptions {
-                    name_override: Some("restorable".to_string()),
-                    ..ApplyOptions::default()
-                },
-            )
-            .expect("first apply should succeed");
-            apply_overlay(
-                overlay_b.path().to_str().unwrap(),
-                ctx.repo_path(),
-                &ApplyOptions {
-                    name_override: Some("missing-source".to_string()),
-                    ..ApplyOptions::default()
-                },
-            )
-            .expect("second apply should succeed");
-
-            fs::remove_dir_all(ctx.repo_path().join(".repoverlay")).unwrap();
-            fs::remove_file(ctx.repo_path().join(".envrc")).unwrap();
-            fs::remove_file(ctx.repo_path().join(".toolrc")).unwrap();
-            fs::remove_dir_all(overlay_b.path()).unwrap();
-
-            let error =
-                restore_overlays(ctx.repo_path(), false, ConflictStrategy::default(), false)
-                    .expect_err("restore should fail when any overlay cannot be restored");
-            let message = error.to_string();
-
-            assert!(
-                ctx.file_exists(".envrc"),
-                "restore should continue and restore successful overlays"
-            );
-            assert!(
-                !ctx.file_exists(".toolrc"),
-                "failed overlay file should remain missing"
-            );
-            assert!(
-                message.contains("restorable"),
-                "summary should include successful overlay name: {message}"
-            );
-            assert!(
-                message.contains("missing-source"),
-                "summary should include failed overlay name: {message}"
-            );
-        }
-
-        #[test]
-        fn reapplying_overlay_clears_removed_marker() {
-            // This test verifies that re-applying an overlay clears the removed_at marker
-            // in case the user changes their mind after removal.
-
-            let ctx = TestContext::new().with_overlay(&[(".envrc", "export FOO=bar")]);
-            let canonical_repo_path = ctx.repo_path().canonicalize().unwrap();
-
-            // Step 1: Apply the overlay
-            apply_overlay(
-                ctx.overlay_source(),
-                ctx.repo_path(),
-                &ApplyOptions::default(),
-            )
-            .expect("apply should succeed");
-
-            let applied = list_applied_overlays(ctx.repo_path()).expect("list should work");
-            let overlay_name = &applied[0];
-
-            // Step 2: Remove the overlay (marks removed_at)
-            remove_overlay(
-                ctx.repo_path(),
-                Some(overlay_name.to_string()),
-                false,
-                false,
-            )
-            .expect("remove should succeed");
-
-            // Verify removed_at is set
-            let ext_dir = external_state_dir_for_target(&canonical_repo_path).unwrap();
-            let ext_state_file = ext_dir.join(format!("{overlay_name}.ccl"));
-            let content = fs::read_to_string(&ext_state_file).unwrap();
-            let ext_state: OverlayState = sickle::from_str(&content).unwrap();
-            assert!(ext_state.removed_at.is_some());
-
-            // Step 3: Re-apply the overlay
-            apply_overlay(
-                ctx.overlay_source(),
-                ctx.repo_path(),
-                &ApplyOptions::default(),
-            )
-            .expect("re-apply should succeed");
-
-            // Step 4: Verify removed_at is cleared
-            let content_after = fs::read_to_string(&ext_state_file).unwrap();
-            let ext_state_after: OverlayState = sickle::from_str(&content_after).unwrap();
-            assert!(
-                ext_state_after.removed_at.is_none(),
-                "removed_at should be cleared after re-apply"
-            );
-
-            // Verify restore would now restore this overlay
-            // (if git clean happened again)
-            let ext_states = load_external_states(&canonical_repo_path).unwrap();
-            assert_eq!(ext_states.len(), 1, "external state should be loadable");
         }
     }
